requirements.txt
Django==1.9.6
django-braces==1.9.0
djangorestframework==3.4.6
mysqlclient==1.3.7
oauthlib==1.0.3
pip==8.1.2
setuptools==26.1.1
six==1.10.0
uWSGI==2.0.13.1
wheel==0.29.0
